pyproject.toml
[project]
name = "lol-chat-helper"
version = "0.1.0"
description = "LOL Chat Helper with memory and OP.GG MCP tools using LangChain and LM Studio"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-core>=0.3.0",
    "langgraph>=0.2.0",
    "langchain-mcp-adapters>=0.1.0",
    "python-dotenv>=1.0.0",
]
